requirements.txt
web3
hdwallet
python-dotenv
mnemonic